Fucking mess

+++ emloop/api.py
@@ -215,11 +215,12 @@ def delete_output_dir(output_dir: str) -> None:
     """
     Delete given output dir.
     """
-    logging.info(f'Delete output dir {output_dir}')
-    shutil.rmtree(output_dir)
+    if path.exists(output_dir):
+        logging.info(f'Delete output dir {output_dir}')
+        shutil.rmtree(output_dir)
 
 
-def create_emloop_training(config: dict, output_root: str, restore_from: str=None) -> MainLoop:
+def create_emloop_training(config: dict, output_root: str, restore_from: str=None) -> EmloopTraining:
     """
     Creates :py:class:`MainLoop` with model, dataset and hooks according to config.
 
@@ -229,6 +230,7 @@ def create_emloop_training(config: dict, output_root: str, restore_from: str=Non
 
     :return: main loop object
     """
+    print('I WAS CALLED ********************8')
     output_dir = dataset = model = hooks = main_loop = None
 
     output_dir = create_output_dir(config=config, output_root=output_root)
